+++ lib/github_stalking/runner.ex
@@ -1,10 +1,10 @@
 defmodule GithubStalking.Runner do
   def run(options) do
     case options do
-      %{:register => repo_full_path} -> 
-        GithubStalking.register_repo(repo_full_path)
+      %{:register => repo_full_path, :collect => nil} -> 
+        GithubStalking.Repository.register_repo(repo_full_path)
 
-      %{:collect => collect} -> 
+      %{:register => nil, :collect => collect} -> 
         #TODO #22
         #GithubStalking.collect_repos_info
         nil
