+++ test/riak_test.exs
@@ -70,15 +70,4 @@ defmodule GithubStalking.RiakTest do
     assert issues_numbers.numbers == Enum.to_list 6..16
   end
 
-  test "register repo" do
-    repo_full_path = "octocat/Spoon-Knife"
-    result = GithubStalking.Riak.register_repo(repo_full_path)
-    assert :ok == result
-  end
-
-  test "register already exists repo" do
-    repo_full_path = "letusfly85/github_stalking"
-    result = GithubStalking.Riak.register_repo(repo_full_path)
-    assert :error == result
-  end
 end
